(FACT-2986) Fix virtual fact on Amazon 6
Due to the additional `check_dmi` method in the VirtualDetector module,
Facter 4 would detect `xen` as hypervisor instead of `xenhvm` on Amazon 6.

This is because in Facter 3, there is no DMI check and virt-what has
higher priority. In our case, the detection stops in the `check_dmi`
method, which returns `xen` and is redundant, as the same functionality
is implemented below in the `check_other_facts` method.

Additionally, the dmidecode resolver doesn't need to add a `vendor`
fact, since the DmiBios resolver already makes it available through the
`bios_vendor` key. So, remove the key from dmidecode, and remove the
`check_dmi` method from VirtualDetector.

The tests were also broken in the `xen-hvm` case, since they were
stubbing the call to `DmiDecode.resolve` to return nil, which in reality
returned 'Xen'. Remove the extra tests and fix some typos.

We should rework these tests with some real life scenarios if possible,
as overstubbing can lead to scenarios like this.