(FACT-2998) Make FactCollection persistent
Previously, when sorting objects of type `FactCollection` using the
`Utils.sort_hash_by_key` method, we would always get a `Hash` back
instead of a `FactCollection`. This was most likely a bug that passed
unnoticed until now.

Among other things, this caused methods that were defined on
`FactCollection` to become unusable.

This commit changes the sort function to maintain the original class,
and adds an encode_with method on `FactCollection` which makes Psych not
add tags and other useless information when calling `#to_yaml` on a
`FactCollection`.