(FACT-2983) Only convert hypervisors to bool
In 9421cf371f2e4d202dbc1c199125eba96c13e67c the Solaris LDOM resolver
was changed to always return boolean values for 'true' or 'false'.
However, Facter 3 only reported bool values for the `hypervisors` fact,
not the `ldom` fact, and the same resolver provides data for both facts.

This broke the `virtual` fact since it specifically checked for a string
value and we returned a boolean value.

To match the output of Facter 3, try to convert only the `hypervisors`
fact to bool.