(FACT-3007) Use VirtualDetector in xen fact
Previously we could end up in a situation where both kvm and xen facts
were resolved. This happens on Amazon 7 where the real hypervisor
(reported by virt-what) is kvm, but the disk devices are still named
`/dev/xvda` which causes xen to also be detected.

Use the value returned by VirtualDetector#platform to determine if we're
running on xen.